+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Server: Support JS packages in importmap.rb and package-lock.json
 - Server: Change launch behavior to more reliably start cache warmer at launch.
 - Server: Add --open option to open the server in a browser on launch.
 - Server: Defer initial detail rendering so the page becomes interactive sooner on startup